subsystem: Split the library into multiple modules
